+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added 8 new achievements!
   - Added Trophy Fishing achievement category
 - Added Lilypad Size Display feature
+- Added a keybind to peek the Party Finder window without opening the GUI
 
 ### Changes
 - Removed 1.21.11 Support